requirements.txt
torch
torchvision
albumentations
pillow
numpy
pydensecrf
matplotlib
pandas
imgaug
Augmentor
opencv-python
solt
